+++ R/h5Utility.R
@@ -43,6 +43,7 @@ H5Apply <- function(
         ...
 ) {
     fun.args <- list(...)
+    da_rowidx <- da_colidx <- NULL
     if (inherits(object, "ligerDataset")) {
         useData <- match.arg(useData)
         h5meta <- h5fileInfo(object)
@@ -51,17 +52,31 @@ H5Apply <- function(
         colptr <- h5file[[h5meta$indptrName]]
         rowind <- h5file[[h5meta$indicesName]]
         data <- h5file[[h5meta[[useData]]]]
+        numCells <- ncol(object)
+        numFeatures <- nrow(object)
+        featureNames <- rownames(object)
     } else if (inherits(object, "DelayedArray")) {
-        h5Path <- get_DelayedArray_filepath(object)
+        h5Path <- DelayedArray::path(object)
         h5file <- hdf5r::H5File$new(h5Path, mode = "r")
         on.exit(h5file$close_all())
-        h5group <- get_DelayedArray_group(object)
+        h5group <- DelayedArray::seed(object)@group
         colptr <- h5file[[file.path(h5group, "indptr")]]
         rowind <- h5file[[file.path(h5group, "indices")]]
         data <- h5file[[file.path(h5group, "data")]]
+        numCells <- h5file[[file.path(h5group, "shape")]][2]
+        numFeatures <- h5file[[file.path(h5group, "shape")]][1]
+        featureNames <- h5file[[file.path(h5group, "features/name")]][]
+        # DelayedArray can be subset/rearranged from the original H5 data
+        da_rowidx <- DelayedArray::netSubsetAndAperm(object)[[1]]
+        da_colidx <- DelayedArray::netSubsetAndAperm(object)[[2]]
+        if (!is.null(da_colidx)) {
+            cli::cli_abort(c(
+                x = "rliger internal H5 chunk computation does not accept DelayedArray representation with column subset or rearrangement.",
+                i = "Please write the subset/rearranged data to new H5 file and insert back to the liger object."
+            ))
+        }
     }
-    numCells <- ncol(object)
-    numFeatures <- nrow(object)
+
 
     prev_end_col <- 1
     prev_end_data <- 1
@@ -71,7 +86,6 @@ H5Apply <- function(
     if (isTRUE(verbose))
         cliID <- cli::cli_progress_bar(name = "HDF5 chunk processing", type = "iter",
                                        total = numChunks, clear = FALSE)
-        # pb <- utils::txtProgressBar(0, numChunks, style = 3)
     for (i in seq(numChunks)) {
         start <- (i - 1)*chunkSize + 1
         end <- if (i*chunkSize > ncol(object)) ncol(object) else i*chunkSize
@@ -88,17 +102,15 @@ H5Apply <- function(
         chunk <- Matrix::sparseMatrix(i = chunkRowind + 1, p = chunkColptr,
                                       x = chunkData,
                                       dims = c(numFeatures, end - start + 1),
-                                      dimnames = list(rownames(object),
+                                      dimnames = list(featureNames,
                                                       colnames(object)[start:end]))
+        if (!is.null(da_rowidx)) chunk <- chunk[da_rowidx, , drop = FALSE]
         init <- do.call(FUN, c(list(chunk, nnzStart:nnzEnd,
                                     start:end, init),
                                fun.args))
-        # if (isTRUE(verbose)) utils::setTxtProgressBar(pb, i)
         if (isTRUE(verbose)) cli::cli_progress_update(id = cliID, set = i)
     }
-    # Break a new line otherwise next message comes right after the "%" sign.
-    if (isTRUE(verbose)) cat("\n")
-    init
+    return(init)
 }
 
 # Safely add new H5 Data to the HDF5 file in a ligerDataset object
@@ -586,3 +598,4 @@ get_DelayedArray_group <- function(x) {
         cli::cli_abort("Cannot detect data path within the HDF5Array-backed DelayedArray.")
     }
 }
+